# backend/src/main/resources/openapi.yml
openapi: 3.1.0
info:
  title: Anveshak API
  version: 1.0.0
  description: API documentation for the Anveshak research paper management application.
servers:
  - url: /
...
security:
  - bearerAuth: []
paths:
...
  /users/me:
    get:
      tags: [User Management]
      operationId: getCurrentUser
      summary: Get the current user's profile
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags: [User Management]
      operationId: updateCurrentUser
      summary: Update the current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/me/password:
    patch:
      tags: [User Management]
      operationId: changePassword
      summary: Change the current user's password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '204':
          description: Password changed
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /users/me/link/google:
    post:
      tags: [User Management]
      operationId: linkGoogleIdentity
      summary: Link a Google identity to the current user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleLoginRequest'
      responses:
        '201':
          description: Google identity linked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserIdentityResponse'
    delete:
      tags: [User Management]
      operationId: unlinkGoogleIdentity
      summary: Unlink the current user's Google identity
      responses:
        '204':
          description: Google identity unlinked
...
  /users/me/sessions/{sessionId}:
    delete:
      tags: [User Management]
      operationId: revokeSession
      summary: Revoke a specific session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Session revoked
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /papers:
    post:
      tags: [Research Paper Management]
      operationId: createPaper
      summary: Create a research paper with a PDF upload
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [paper, pdfFile]
              properties:
                paper:
                  $ref: '#/components/schemas/NewPaperRequest'
                pdfFile:
                  type: string
                  format: binary
            encoding:
              paper:
                contentType: application/json
      responses:
        '201':
          description: Paper created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchPaperResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags: [Research Paper Management]
      operationId: listPapers
      summary: List research papers for the current user
      responses:
        '200':
          description: Papers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResearchPaperResponse'
  /papers/{paperId}:
    get:
      tags: [Research Paper Management]
      operationId: getPaper
      summary: Get a research paper by ID
      parameters:
        - name: paperId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Paper details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchPaperResponse'
    patch:
      tags: [Research Paper Management]
      operationId: updatePaper
      summary: Update a research paper and optionally replace its PDF
      parameters:
        - name: paperId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [paper]
              properties:
                paper:
                  $ref: '#/components/schemas/UpdatePaperRequest'
                pdfFile:
                  type: string
                  format: binary
            encoding:
              paper:
                contentType: application/json
      responses:
        '200':
          description: Paper updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchPaperResponse'
    delete:
      tags: [Research Paper Management]
      operationId: deletePaper
      summary: Delete a research paper
      parameters:
        - name: paperId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Paper deleted
  /papers/{paperId}/pdf:
    get:
      tags: [Research Paper Management]
      operationId: downloadPaperPdf
      summary: Download the paper PDF
      parameters:
        - name: paperId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: PDF file
          content:
            application/pdf:
              schema:
                type: string
                format: binary
  /papers/search:
    get:
      tags: [Research Paper Management]
      operationId: searchPapers
      summary: Search papers by query
      parameters:
        - name: query
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Matching papers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResearchPaperResponse'
  /papers/compare:
    post:
      tags: [Research Paper Management]
      operationId: comparePapers
      summary: Compare papers by ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaperComparisonRequest'
      responses:
        '200':
          description: Comparison result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaperComparisonResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: jwt
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        code:
          type: integer
...
    GoogleLoginRequest:
      type: object
      properties:
        token_id:
          type: string
          description: Google ID token.
...
    UserProfileResponse:
      type: object
      required: [id, firstName, lastName, username, email, emailVerified, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
        emailVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UpdateProfileRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        username:
          type: string
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
    UserIdentityResponse:
      type: object
      required: [id, provider, providerUserId, createdAt]
      properties:
        id:
          type: string
          format: uuid
        provider:
          type: string
        providerUserId:
          type: string
        createdAt:
          type: string
          format: date-time
...
    NewPaperRequest:
      type: object
      required: [title, abstractText, authors, publicationYear, keywords]
      properties:
        title:
          type: string
        abstractText:
          type: string
        authors:
          type: array
          items:
            type: string
        publicationYear:
          type: integer
        keywords:
          type: array
          items:
            type: string
    UpdatePaperRequest:
      type: object
      properties:
        title:
          type: string
        abstractText:
          type: string
        authors:
          type: array
          items:
            type: string
        publicationYear:
          type: integer
        keywords:
          type: array
          items:
            type: string
    ResearchPaperResponse:
      type: object
      required: [id, title, abstractText, authors, keywords, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        abstractText:
          type: string
        authors:
          type: array
          items:
            type: string
        keywords:
          type: array
          items:
            type: string
        publicationYear:
          type: integer
          nullable: true
        pdfUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PaperComparisonRequest:
      type: object
      required: [paperIdStrings]
      properties:
        paperIdStrings:
          type: array
          items:
            type: string
    PaperComparisonResponse:
      type: object
      required: [papers]
      properties:
        papers:
          type: array
          items:
            $ref: '#/components/schemas/PaperComparison'
    PaperComparison:
      type: object
      required: [title, objective, methodology, dataset, results, strengths, weaknesses, futureWork]
      properties:
        title:
          type: string
        objective:
          type: string
        methodology:
          type: string
        dataset:
          type: string
        results:
          type: string
        strengths:
          type: string
        weaknesses:
          type: string
        futureWork:
          type: string